MS1/requirements.txt
django
djangorestframework
psycopg2-binary
gunicorn
django-heroku
django-rest-auth
django-allauth
django-rest-auth[with_social]
